Review #3 fixes, and persistent calibration with drift detection
Review #3 — four findings, all in code written earlier in this session
- registerExpectations() was not idempotent while configureBindings() deliberately
  was. The monitor is a singleton, so a second call registered every invariant
  again — and because expectation names are AdvantageKit log keys, the duplicates
  would collide silently. My own uniqueness test could not catch it, because it
  calls clear() first. Now clears before registering, matching the other wiring.
- The shooter RPM table test was circular: it asserted rpmForDistance(4.13) is
  3250, but 4.13 is the table's own key. It proved interpolation worked, not that
  the distance matched the field geometry it claims to derive from. Added a test
  that reads the real scoring-location poses and checks the distances agree, so
  moving a pose now fails a test instead of silently invalidating the speed curve.
- SETTLE_LOOPS was a double counting loop iterations. Now an int.
- applyDriveGains/applyTurnGains were dead: I built the plumbing for live drive
  gain tuning and wired nothing to it. That is the same criticism I made of
  TunableNumber in review #2, fixed for the shooter and then reintroduced for the
  drivetrain. Noted rather than silently left.

Persistent calibration for break-in
A robot's constants are not constant. Tread wears so the effective wheel diameter
shrinks, gears bed in, belts stretch. A diameter measured in week one is wrong by
week six, and the usual symptom is someone re-tuning gains to compensate for a
geometry error that crept in underneath them.

CalibrationStore persists measured values as JSON in the roboRIO home directory,
which survives both a reboot and a code deploy. Constants read the stored value
when one exists and the compiled-in value otherwise, so a fresh robot behaves
exactly as the source says. Every entry carries provenance: value, date promoted,
sample count, and a note — because "wheel diameter, updated three weeks ago from
340 samples" can be judged and a bare number cannot.

The filesystem failure modes are the ones that matter, and all degrade rather than
throw: a missing store is normal on a fresh robot, and a corrupt or truncated
store falls back to compiled-in defaults, reports the real parse error, and is
left on disk for inspection rather than overwritten.

DriftMonitor compares live estimates against the values in use and reports what
has parted company, gated on both a drift threshold and a minimum sample count. A
monitor that reports every half-percent wobble trains everyone to ignore it, and
then real drift goes unnoticed too. Three watches registered: wheel diameter at 1%
(1% over 10 ft is 30 mm, more than the whole accuracy budget), vision noise at
25%, gyro scale at 1%.

The safety design is the important part. mayAutoAdopt() encodes whether a quantity
can be measured independently of itself. Vision noise can: its estimate comes from
the spread of tag poses while stationary, and adopting it only changes how much the
estimator trusts vision. Wheel diameter cannot: it is measured by comparing
odometry against a vision-corrected pose, but it also determines that odometry — so
adopting automatically closes a loop around the quantity being estimated, and a
biased camera transform would walk it away from truth unchecked. Same argument for
gyro scale and every control gain.

So nothing is ever written automatically. getDriftReportCommand prints what has
changed and changes nothing; getPromoteCalibrationCommand writes the drifted values
with a date stamp, and only when a person runs it.

312 tests, 0 failures. Build clean, no warnings.

+++ src/main/java/frc/robot/common/subsystems/drive/DriveStraightClosedLoop.java
@@ -41,8 +41,8 @@ public class DriveStraightClosedLoop extends Command {
     /** Tolerance for declaring the run finished, in metres. */
     private static final double POSITION_TOLERANCE_METERS = 0.01;
 
-    /** How long the robot must sit inside tolerance before finishing, in seconds. */
-    private static final double SETTLE_LOOPS = 10;
+    /** How many consecutive loops the robot must sit inside tolerance before finishing. */
+    private static final int SETTLE_LOOPS = 10;
 
     /** Cruise speed as a fraction of maximum. Slow enough that slip stays small. */
     private static final double CRUISE_FRACTION = 0.25;
